# (NFC) RegionTest, SmartyTest - Restore performance