--- a/tests/phpunit/CRM/Core/Smarty/plugins/CrmPermissionTest.php
+++ b/tests/phpunit/CRM/Core/Smarty/plugins/CrmPermissionTest.php
@@ -7,6 +7,11 @@
  */
 class CRM_Core_Smarty_plugins_CrmPermissionTest extends CiviUnitTestCase {
 
+  public function setUp(): void {
+    parent::setUp();
+    $this->useTransaction();
+  }
+
   /**
    * @dataProvider permissionCases
    *