--- a/tests/phpunit/CRM/Core/Smarty/plugins/CrmScopeTest.php
+++ b/tests/phpunit/CRM/Core/Smarty/plugins/CrmScopeTest.php
@@ -6,6 +6,11 @@
  */
 class CRM_Core_Smarty_plugins_CrmScopeTest extends CiviUnitTestCase {
 
+  public function setUp(): void {
+    parent::setUp();
+    $this->useTransaction();
+  }
+
   /**
    * @return array
    */