--- a/tests/phpunit/CRM/Core/Smarty/plugins/UrlTest.php
+++ b/tests/phpunit/CRM/Core/Smarty/plugins/UrlTest.php
@@ -6,6 +6,11 @@
  */
 class CRM_Core_Smarty_plugins_UrlTest extends CiviUnitTestCase {
 
+  public function setUp(): void {
+    parent::setUp();
+    $this->useTransaction();
+  }
+
   /**
    * @return array
    */