Overview
----------------------------------------

This appears to have been collateral damage from e9cceb3799e74357c96123a0456a27c5b5468856.

Before + After
----------------------------------------

| Suite | Runtime (Before) | Runtime (After) | Change |
| -- | -- | -- | -- |
| `tests/phpunit/CRM/Core/RegionTest.php` | 57 seconds | 30 seconds | -27 seconds |
| `tests/phpunit/CRM/Core/Smarty/` | 79 seconds | 55 seconds | -24 seconds |

Overall, this cuts a minute from the CRM test-suite.

(*Times measured on Intel 12600k. Values are rounded and may vary slightly but were generally consistent in a couple runs.*)